video wrapper

+++ src/video_manager/downloaded_videos.hpp
@@ -6,7 +6,7 @@
 
 class DownloadedVideos {
 public:
-    std::vector<int> downloaded_ids;
+    std::vector<int64_t> downloaded_ids;
 
     DownloadedVideos() = default;
     DownloadedVideos(const std::vector<int> &ids);
